+++ asset_wrapper/__init__.py
@@ -4,7 +4,7 @@
 bl_info = {
     "name": "Asset Wrapper",
     "author": "Blue Moon Virtual",
-    "version": (0, 6, 0),
+    "version": (0, 7, 0),
     "blender": (4, 2, 0),
     "location": "View3D > Sidebar > Asset Wrapper",
     "description": (
@@ -27,6 +27,8 @@ classes = (
     operators.AW_OT_refresh_library,
     operators.AW_OT_open_library_folder,
     operators.AW_OT_remove_asset_file,
+    operators.AW_OT_edit_asset,
+    operators.AW_OT_reload_libraries,
     operators.AW_OT_rename_asset,
     operators.AW_OT_batch_rename,
     operators.AW_OT_set_custom_folder,